Move animation into another component
I should do this with real animation, but I don't want to dive into that
right now.

Being its own component makes it easier to cache positions so I can get
back to the original position and orientation.

Also update the prefab to only Land attack at the end of the animation.
I'll need to fix this so we just keep playing animation, but are "Done".